% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{check_dependency}
\alias{check_dependency}
\title{Title Check SurvDeepFIT's dependency, especially for DeepSurv and DeepHit}
\usage{
check_dependency(package = "SurvDeepFIT", use_condaenv = "r-reticulate")
}
\arguments{
\item{package}{The name of the package to check dependencies for. Default is "SurvDeepFIT".}

\item{use_condaenv}{The name of the Conda environment to use, as specified by reticulate. Can be changed to a specific Conda environment.}
}
\value{
This function checks the dependencies for SurvDeepFIT, DeepSurv, DeepHit, and other related machine learning methods.
}
\description{
Title Check SurvDeepFIT's dependency, especially for DeepSurv and DeepHit
}
\examples{
check_dependency()
}
\author{
Shiyu Wan
}
