% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{model.centralize}
\alias{model.centralize}
\title{Title Centralizing the Relative Risk Prediction of SurvDeepFIT Around 0}
\usage{
model.centralize(mod, object)
}
\arguments{
\item{mod}{A fitted dnnet or dnnetEnsemble object.}

\item{object}{The training set corresponding to the above model.}
}
\value{
A centralized model with the last bias updated, and the event times of the training set.
}
\description{
Title Centralizing the Relative Risk Prediction of SurvDeepFIT Around 0
}
