% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{timedAUC}
\alias{timedAUC}
\title{Title Internal Method to Calculate Time-Dependent AUC}
\usage{
timedAUC(Status, Times, f_hat_y)
}
\arguments{
\item{Status}{A numeric vector of patients' survival status: 1 = event, 0 = censored.}

\item{Times}{A numeric vector of patients' survival times.}

\item{f_hat_y}{A numeric vector of patients' survival risk predictions.}
}
\value{
The average of time-dependent AUCs at the 25\%, 50\%, and 75\% quantiles of the event time.
}
\description{
Title Internal Method to Calculate Time-Dependent AUC
}
