% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{check_dependency}
\alias{check_dependency}
\title{Check PermSurvDNN's dependency}
\usage{
check_dependency(package = "PermSurvDNN", use_condaenv = "r-reticulate")
}
\arguments{
\item{package}{Default Package Name: PermSurvDNN}

\item{use_condaenv}{Default Conda Environment's name given by reticulate; can be changed to specific conda environment}
}
\value{
This function is used to check the dependency of PermSurvDNN, DeepSurv and other machine learning methods.
}
\description{
Check PermSurvDNN's dependency
}
\examples{
check_dependency()
}
\author{
Shiyu Wan
}
