% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{check_dependency}
\alias{check_dependency}
\title{Check PermSurvDNN's dependency}
\usage{
check_dependency(package = "PermSurvDNN")
}
\arguments{
\item{package}{}
}
\value{
Check Dependency
}
\description{
Check PermSurvDNN's dependency
}
\examples{
TBD
}
\author{
Shiyu Wan
}
