% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{log_lik_diff}
\alias{log_lik_diff}
\title{Title  Internal Method for Measuring Prediction Performance Differences in Survival Models}
\usage{
log_lik_diff(
  model.type,
  y_hat,
  y_hat0,
  object,
  y_max = 1 - 10^-10,
  y_min = 10^-10,
  y_hatcoxl,
  y_hat0coxl
)
}
\arguments{
\item{model.type}{Default = "survival" for survival data.}

\item{y_hat}{A numeric vector of patients' survival risk prediction.}

\item{y_hat0}{A numeric vector of patients' survival risk prediction based on permuted data (Details explained in the manuscript).}

\item{object}{Validation dataset, a dnnetSurvInput object, created by deepTL::importDnnetSurv().}

\item{y_max}{Inner upper boundary for y in binary classification, can be ignored when model.type = "survival".}

\item{y_min}{Inner lower boundary for y in binary classification, can be ignored when model.type = "survival".}

\item{y_hatcoxl}{A numeric vector of patients' survival risk prediction returned by Cox-related models, e.g. Cox, SurvDNN, XGBoost.}

\item{y_hat0coxl}{A numeric vector of patients' survival risk prediction returned by Cox-related models based on permuted data.}
}
\value{
A numeric vector consisting of differences in C-index, Cox's partial log-likelihood, and average time-dependent AUC, based on observed and permuted data.
}
\description{
Title  Internal Method for Measuring Prediction Performance Differences in Survival Models
}
\examples{
TBD
}
