% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{loglik_coxph}
\alias{loglik_coxph}
\title{Calculated the Cox-PH's log-likelihood based on event status, event time and risk prediction}
\usage{
loglik_coxph(Status, Times, f_hat_y)
}
\arguments{
\item{Status}{Patients' event statuses}

\item{Times}{Patients' event times}

\item{f_hat_y}{X \%*\% Beta}
}
\value{
Returns the log-likelihood of a Cox-PH Model
}
\description{
Calculated the Cox-PH's log-likelihood based on event status, event time and risk prediction
}
