% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{model.centralize}
\alias{model.centralize}
\title{Title Centralizing the relative risk prediction of SurvDNN around 0}
\usage{
model.centralize(mod, object)
}
\arguments{
\item{mod}{A fitted dnnet or dnnetEnsemble object}

\item{object}{The training set corresponds to the above model}
}
\value{
A centralized model with the last bias being updated, and the event time of the Training set
}
\description{
Title Centralizing the relative risk prediction of SurvDNN around 0
}
