% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{predict_surv_df}
\alias{predict_surv_df}
\title{Title Internal Method to predict patient's survival based on relative risk prediction and Breslow's estimator of baseline cumulative hazard}
\usage{
predict_surv_df(centralized_model, test_object, time_point = NULL)
}
\arguments{
\item{centralized_model}{The output of function model.centralize().}

\item{test_object}{Test dataset, a dnnetSurvInput object}

\item{time_point}{If specified, then it corresponds to the specific time points where you want to predict the patient's survival; otherwise, survival probability of each unique failure time in the training dataset willbe provided}
}
\value{
A data.frame with each patient's survival probability prediction.
}
\description{
Title Internal Method to predict patient's survival based on relative risk prediction and Breslow's estimator of baseline cumulative hazard
}
