% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_permfit_survival_20230610.r
\name{timedAUC}
\alias{timedAUC}
\title{Title Internal Method to Calculate Time-dependent AUC}
\usage{
timedAUC(Status, Times, f_hat_y)
}
\arguments{
\item{Status}{A numeric vector of patients' survival status: 1 = event and 0 = censor}

\item{Times}{A numeric vector of patients' survival times}

\item{f_hat_y}{A numeric vector of patients' survival risk prediction.}
}
\value{
An average of time-dependent AUCs at 25\%, 50\% amd 75\% quantile of the event time.
}
\description{
Title Internal Method to Calculate Time-dependent AUC
}
\examples{
TBD
}
